+++ lib/better_translate.rb
@@ -12,45 +12,47 @@ require 'better_translate/providers/base_provider'
 require 'better_translate/providers/chatgpt_provider'
 require 'better_translate/providers/gemini_provider'
 
+require 'ostruct'
+
 module BetterTranslate
   class << self
     attr_accessor :configuration
 
-    # Metodo per configurare la gemma
+    # Method to configure the gem
     def configure
       self.configuration ||= OpenStruct.new
       yield(configuration) if block_given?
     end
 
-    # Metodo install per generare il file di configurazione (initializer)
+    # Install method to generate the configuration file (initializer)
     def install
       unless defined?(Rails) && Rails.respond_to?(:root)
-        message = "Il metodo install è disponibile solo in un'applicazione Rails."
+        message = "The install method is only available in a Rails application."
         BetterTranslate::Utils.logger(message: message)
         return
       end
 
-      # Costruisce il percorso della cartella template all'interno della gemma
+      # Builds the path to the template folder inside the gem
       source = File.expand_path("../generators/better_translate/templates/better_translate.rb", __dir__)
       destination = File.join(Rails.root, "config", "initializers", "better_translate.rb")
 
       if File.exist?(destination)
-        message = "Il file initializer esiste già: #{destination}"
+        message = "The initializer file already exists: #{destination}"
         BetterTranslate::Utils.logger(message: message)
       else
         FileUtils.mkdir_p(File.dirname(destination))
         FileUtils.cp(source, destination)
-        message = "Il file initializer esiste già: #{destination}"
+        message = "The initializer file already exists: #{destination}"
         BetterTranslate::Utils.logger(message: message)
       end
     end
 
     def magic
-      message = "Metodo magic invocato: eseguirò la traduzione dei file..."
+      message = "Magic method invoked: Translation will begin..."
       BetterTranslate::Utils.logger(message: message)
 
       BetterTranslate::Translator.work
-      message = "Metodo magic invocato: Traduzione completata con successo!"
+      message = "Magic method invoked: Translation completed successfully!"
       BetterTranslate::Utils.logger(message: message)
     end