+++ lib/generators/better_translate/install_generator.rb
@@ -4,7 +4,7 @@ module BetterTranslate
   module Generators
     class InstallGenerator < Rails::Generators::Base
       source_root File.expand_path('templates', __dir__)
-      desc "Genera il file config/initializers/better_translate.rb con la configurazione di default"
+      desc "Generates the config/initializers/better_translate.rb file with default configuration"
 
       def copy_initializer
         template "better_translate.rb", "config/initializers/better_translate.rb"