+++ spec/better_translate_spec.rb
@@ -5,7 +5,15 @@ RSpec.describe BetterTranslate do
     expect(BetterTranslate::VERSION).not_to be nil
   end
 
-  it "does something useful" do
-    expect(false).to eq(true)
+  describe '.configure' do
+    it 'allows configuration of provider and API keys' do
+      BetterTranslate.configure do |config|
+        config.provider = :chatgpt
+        config.openai_key = 'test_key'
+      end
+
+      expect(BetterTranslate.configuration.provider).to eq(:chatgpt)
+      expect(BetterTranslate.configuration.openai_key).to eq('test_key')
+    end
   end
 end